[get-reststops-by-highways] Brand package refactoring
