+++ client/src/main/kotlin/com/eightpotatoes/nexters/client/brand/response/BrandOrigins.kt
@@ -1,4 +1,4 @@
-package com.eightpotatoes.nexters.client.brand
+package com.eightpotatoes.nexters.client.brand.response
 
 data class BrandOrigins(
     val list: List<BrandOrigin>,
